lab1: Будник Михаил - touchups

+++ lab1/gr9sub2/БУДНИК_МИХАИЛ/src/scripts/c.sh
@@ -1,3 +1,8 @@
+if [ ! -f "/var/log/dpkg.log" ]; then
+        echo "файл /var/log/dpkg.log не найден!"
+        exit 1
+fi
+
 # читаем весь файл и ищем появление ивентов установки пакетов
 grep 'install ' /var/log/dpkg.log |
 	# извлекаем название пакета
